+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Clicking on a contact marker in the contacts activity doesn't always center the map on the marker (#1349, #1280)
 - Don't show an error message on non-error conditions exporting config (#1280)
 - Both cloud and local backup should work now
+- Config export actually exports to a local file now, rather than just a somewhat useless "share"
 
 ## Version 2.4.10